moved bandpass to better location
